+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-##[vNext]
+## [vNext]
 ### Added
 - Support cross origin iframes in UFG [Trello 551](https://trello.com/c/iJKPvd75)
 - Add x-applitools-eyes-client-request-id header to API requests [Trello 2332](https://trello.com/c/yuBGjW6u)
+- Add padding option to match regions [Trello 2337](https://trello.com/c/RNNcgMv7)
 ### Updated
 - Pin keyring dependency for travis deploy stage. Update some tests [Trello 2309](https://trello.com/c/aIWIFRcY)